Planning ACI V2.0
- Ajout de la modification d'un planning
- Fix de sécurités
- Optimisation du code et des requêtes
- Améliorations CSS
- Sauvegarder un planning sans adhérents dans chaque équipe est
désormais impossible
- Les dates sont gérées plus proprement & logiquement

+++ v3.0/planning/del_planning.php
@@ -4,26 +4,17 @@ $pwd='../';
 ?>
 <div id="corps">
   <?php
-  	if(isset($_POST['Date']))
+  	if(isset($_POST['Date']) && isset($_POST["tableName"]))
   	{
   		$date=DateTime::createFromFormat('d/m/Y', $_POST["Date"])->format('Y-m-d');
-  		if(isset($_POST["tableName"]))
-  		{
-  			$tableName=$_POST["tableName"];
-			$query = mysqli_query($db, 'DELETE FROM '.$tableName.' WHERE ASSOC_date="'.$date.'";');
-	  		if($query)
-			{
-		        echo '<div id="good">     
-		            <label>Le planning de la semaine du lundi '.$date.' supprimé avec succès !</label>
-		            </div>';
-	        }
-			else
-			{
-				echo '<div id="bad">     
-				      <label>Une erreur s\'est produite lors de la suppression du planning !</label>
-				      </div>';
-			}
-  		}
+		$tableName=$_POST["tableName"];
+		$query = mysqli_query($db, 'DELETE FROM '.$tableName.' WHERE ASSOC_date="'.$date.'";');
+  		if($query)
+		{
+	        echo '<div id="good">     
+	            <label>Le planning de la semaine du lundi '.date("d/m/Y", strtotime($date)).' supprimé avec succès !</label>
+	            </div>';
+        }
 		else
 		{
 			echo '<div id="bad">     
@@ -37,11 +28,12 @@ $pwd='../';
 		      <label>Une erreur s\'est produite lors de la suppression du planning !</label>
 		      </div>';
 	}
+
     echo '<script type="text/javascript">
-	window.location.replace("./homePlanning.php");
-	</script>'
+	window.setTimeout("location=(\'./planning_insertion.php\');",2500);
+	</script>';
   ?>
 </div>
-<?php  
-  include('../footer.php');
+<?php
+	include('../footer.php');
 ?>
